Fixed SonarCloud issue after adding unit test to verify result

+++ src/raspberrypi/devices/ctapdriver.cpp
@@ -452,7 +452,7 @@ uint32_t CTapDriver::Crc32(const BYTE *buf, int length) {
    for (int i = 0; i < length; i++) {
       crc ^= buf[i];
       for (int j = 0; j < 8; j++) {
-         uint32_t mask = -(crc & 1);
+         uint32_t mask = -((int)crc & 1);
          crc = (crc >> 1) ^ (0xEDB88320 & mask);
       }
    }
